# new face detector using OpenCV combined with FaceMesh