… scripts/haarcascade_frontalface_default.xml