![00021](https://user-images.githubusercontent.com/7525026/215374261-ff52ed25-a65a-440b-b518-8a0037425726.jpg)
![00029](https://user-images.githubusercontent.com/7525026/215374365-6aa84db2-d0e9-473f-88c0-eb28d6232449.jpg)

The goal is to make a face detector that detects smaller faces without having to specify subdivision into multiple subimages, so it can just work automatically when processing batches.

To do this, I use OpenCV2's face detector to find rectangles that might be faces, then use the existing FaceMesh code to confirm that the face is real and to get the landmarks to be able to get a mask from it.

Here are the results I get in this version (test X is 32 images, test Y is 176 images, time is number of seconds to detect the faces using Mask Only mode):

|method|faces X|time X|faces Y|time Y|
|--|--|--|--|--|
|original (FaceMesh), 1x1 subimages |30|1.0|136|6.3|
|original (FaceMesh), 3x3 subimages |166|3.5|1063|15.8|
|original (FaceMesh), 5x5 subimages |262|6.2|1849|36.3|
|new (OpenCV) Normal, 30pix minimum |282|4.8|2273|34.0|
|new (OpenCV) Slow, 30pix minimum   |333|23.4|2624|171.6|
|new (OpenCV) Extremely Slow, 30pix minimum|370|130.8|2802|1080.5|

I made a web page showing the masks generated for the 32 images, [it's here](http://nothings.org/ai/face-detect). None of these images were chosen specifically to break the old detector, but I didn't test anything with fewer than 3 faces so obviously that affects the results; the numbers don't reflect the easy cases where both will do equally well.

Some specific cases to notice:
- OpenCV always gets cases like 01-04.
- Both face detectors recognize faces of people wearing masks in 08 and 14! I wasn't expecting this, but figured I'd test it.
- 11 shows how it can be hard to find a setting with the old detector that finds all the faces.
- 12 is an example where the 5th face (further back) is only found by Extremely Slow.
- 15, 16, 23, and 26 are ones where the old detector finds more faces than the new one, but only if you find the right subdivision. I think OpenCV can find it with an even slower mode, which maybe is possible if I optimize (see below).
- For 24, OpenCV never finds the faces at all even though FaceMesh can find them. There's a separate "face in profile" detector that might address this, I'll play with that after this PR goes through.
- 25 is a weird tie with Extremely Slow (both approaches find 5 faces, but not the same ones).

You can see sometimes with 3x3 and 5x5 the face is detected, but it was cropped so the mask has a straight line across it. It might still be okay, I didn't actually test face-swapping those specifically.

More details about how the new detector works:
- First it runs the old detector. OpenCV sometimes misses large faces unless you set it to run even slower than Extremely Slow.
- Next it runs the OpenCV face detector to collect rectangles that might be faces.
- Then it crops the image around each face. FaceMesh only likes faces that are a certain size. Some thorough testing gave me the best results by using a crop that was 4x the face rectangle width, and 2.5x the face rectangle height.
- It gives the cropped image to FaceMesh, which returns (potentially) multiple faces.
- It takes the face landmarks from FaceMesh, finds their convex hulls, computes the bounding boxes, and then looks for a bounding box that overlaps significantly with the OpenCV face rectangle that corresponds to this specific crop.
- If it doesn't find a match with FaceMesh, it adds one to a `rejected` counter which we print if in development mode (see below)

Some things to know / issues to consider:
- There's a new dropdown to select which mode to use. It defaults to the old mode. I'd prefer it to default to the fastest new mode, which I think is a good compromise, but that's up to you. Change `DEFAULT` in face_detect.py.
- I moved all the face detection logic into face_detect.py since there's a lot of code.
- All the subdivision code is still there and still working. If you think the new thing is good enough you might want to get rid of all that code and simplify the UI, but maybe it's better to keep it for people who like the old way. Up to you!
- There's an accompanying slider to set the minimum face size, because this is a parameter for OpenCV's face detector, and it seemed useful to let people say they don't want faces "in the background" detected.
- There's a debug UI controlled by the setting of `FaceDetectDevelopment` in face_detect.py
- I pass a new variable 'facecfg' all around lots of functions. It's primarily used to pass the debug UI settings. I'm not sure if the normal convention is to just put stuff on `p` if you need some kind of context like this, but it also needs to work if called from UpdateVisualizer so it's kind of messy.
- I had a bunch of messy trailing whitespaces in my new code so I used search-and-replace to clean that all up, but it also cleaned it up from your code, which makes the diff a bit messier, sorry.
- I hacked in a `faceDebug` function to save those masked images found on the web page (you have to turn on FaceDetectDevelopment to use it). It uses a hacky image save routine called "debugsave".  I didn't know the best way to specify an output directory for that, so if you want to change it to something saner (I use an environment variable) that's fine.
- None of this is tuned as well as I'd like. I got distracted by trying to optimize instead.
- There's a bunch of optimizations I tried but they're all disabled (at the end of face_detect.py). Some should be viable to make Extremely Slow faster, but I didn't fully understand the results I was getting, so I need more time to figure out what's really going on.
- Some stuff is heavily commented, some stuff isn't.
- This is my first time programming Python. I did a lot of googling to try to find the most Pythonic ways to do things, but there be some dumb stuff still.

I'm happy to make any changes you ask for!