@nothings 2023-01-30
Also, in my testing, sometimes FaceMesh alone could find 5 images, so I bumped the setting in the FaceMesh call up to 5 even in the old detector path. Could maybe make it 6 just in case.