@kex0 2023-01-30
This is beyond amazing. You've put some serious work into this and I really appreciate it.

I didn't read through the code yet but from my very limited testing, it works very good. I'll do some more extensive testing tomorrow, I have to get some sleep.

Nonetheless, I added you as a contributor to this repository, so feel free to merge this and add any changes you want.

I'm astonished from the amount and the quality of the work you put into this.

P.S. the "haarcascade_frontalface_default.xml" has to be included in the repository.