@nothings 2023-01-30
I'll let you test it first, it's definitely possible I broke something since I didn't test all possible paths (I'm not sure I even know what they are). Once you've tested, you can push it or I can push it, either one is fine. Do read over the list at the end at the end of the submission there are things you'd like changed.

For future changes, I'll probably make a "dev" branch and push stuff that I work on to that and ask if you have any opinions on it, then we can push it to main afterwards.