@nothings 2023-01-30
It looks like the OpenCV detector is CPU not GPU, and it uses multiple CPUs. My machine has 16 cores, so results may be a lot slower on machines with fewer. Might be worth testing if you have a very different setup. (Most important thing to compare is 5x5 old detector vs. Normal new detector.)

But then again, maybe FaceMesh is the same, so it works out, I don't know.