@kex0 2023-01-30
> * I'd prefer it to default to the fastest new mode, which I think is a good compromise,

I agree.

> * All the subdivision code is still there and still working. If you think the new thing is good enough you might want to get rid of all that code and simplify the UI, but maybe it's better to keep it for people who like the old way. Up to you!

Do I understand it correctly that it first runs mediapipe, then OpenCV, crops the image and runs mediapipe on the cropped image? If so then all the subdivision code is pretty much useless considering that the new fastest way is much more reliable and  slightly faster than 5x5 subdivision.

It needlessly adds complexity for users so I think it can be completely removed.

> * This is my first time programming Python. I did a lot of googling to try to find the most Pythonic ways to do things, but there be some dumb stuff still.

I'm impressed that you managed to make some sense of my complete mess of a code. I'm sure you can tell I'm a beginner myself.

> I'll let you test it first, it's definitely possible I broke something since I didn't test all possible paths (I'm not sure I even know what they are). Once you've tested, you can push it or I can push it, either one is fine. Do read over the list at the end at the end of the submission there are things you'd like changed.

I tested everything I could think of and the new detection seems to work very well. I made some tiny UI changes but that's it. (not sure how to commit to this pull request, I'll try)

> For future changes, I'll probably make a "dev" branch and push stuff that I work on to that and ask if you have any opinions on it, then we can push it to main afterwards.

Good idea.

> It looks like the OpenCV detector is CPU not GPU, and it takes advantage of multiple CPUs. My machine has 16 cores, so results may be a lot slower on machines with fewer. Might be worth testing if you have a very different setup. (Most important thing to compare is 5x5 old detector vs. Normal new detector.)
> 
> But then again, maybe FaceMesh is the same, so it works out, I don't know.

Yeah, FaceMesh also uses CPU. My CPU is also quite beefy and I don't have access to anything much slower so I can't test it.