@nothings 2023-01-31
> > * All the subdivision code is still there and still working. If you think the new thing is good enough you might want to get rid of all that code and simplify the UI, but maybe it's better to keep it for people who like the old way. Up to you!
> 
> Do I understand it correctly that it first runs mediapipe, then OpenCV, crops the image and runs mediapipe on the cropped image? If so then all the subdivision code is pretty much useless considering that the new fastest way is much more reliable and slightly faster than 5x5 subdivision.

That understanding is correct; the new detector still runs FaceMesh once for every potential face, just with a custom crop instead of an arbitrary subimage, so it's meant to replace the subdivision approach.

The only thing is that it's not 100% more reliable, as there are a few cases I listed where OpenCV didn't find a face that subdivided FaceMesh did. The new one always finds more faces overall in a large batch, but if you're working on a single image you could tweak the subdivision and sometimes, rarely, get more faces.

Personally I'm fine with getting rid of the subdivision complexity and losing some faces (and gaining more in return), and maybe over time I can address this and improve the results from the OpenCV detector with optimizations and tuning so it also finds those missing faces.