@kex0 2023-01-31
If we were to keep the subdivision code, maybe we could add [THIS](http://nothings.org/ai/face-detect/) to the README or to one of the tips in the webUI itself so users can more easily understand what's happening and how different detectors can affect the speed and the amount of detections.