@nothings 2023-01-31
Yeah, regardless we should definitely update the wiki, I don't think we need that full demo, but I'll definitely writeup some explanation of what's going on. I'm gonna let you decide whether to keep the subdivision code or not, but either way I'll write something up this weekend (if it's ok to wait that long).