@kex0 2023-01-31
> Yeah, regardless we should definitely update the wiki, I don't think we need that full demo,

Agreed. 

> I'm gonna let you decide whether to keep the subdivision code or not,

I think I will keep the subdivision code and just hide it in the UI if the old detection method isn't selected.

> either way I'll write something up this weekend (if it's ok to wait that long).

Sounds great. Of course it's ok.