@nothings 2023-01-31
I just pushed a chance so the UI defaults to using the fastest OpenCV mode.